refactored questions to be one question component and one question bank
