+++ src/routes/routesList.js
@@ -5,7 +5,6 @@ const Home = lazy(() => import("pages/Home/Home"));
 const About = lazy(() => import("pages/About/About"));
 const UserProfile = lazy(() => import("pages/UserProfile/UserProfile"));
 const VibeCheck = lazy(() => import("pages/VibeCheck/VibeCheck"));
-const Results = lazy(() => import("pages/VibeCheck/Results/Results"));
 
 const routes = [
   {
@@ -27,11 +26,6 @@ const routes = [
     path: "/vibecheck",
     exact: true,
     component: VibeCheck
-  },
-  {
-    path: "/results",
-    exact: true,
-    component: Results
   },
   {
     path: "*",
